requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
cors==1.0.1
python-dotenv
pymysql